Use Kerberos Test Harness for Unit Tests
Previously, it was necessary to have a fully functioning
Kerberos setup with a 'host' service for the unit tests to
run properly.  Now they instead use a heavily modified version
of the MIT Kerberos test harness (k5test.py), so only Kerberos
binaries and libararies are needed, not a working installation.

This has the added benefit of bein self-contained.